+++ src/main/java/com/yam/app/account/domain/AccountRepository.java
@@ -7,4 +7,6 @@ public interface AccountRepository {
     boolean existsByNickname(String nickname);
 
     Account save(Account entity);
+
+    Account update(Account entity);
 }
